+++ src/main/java/org/altadoon/gt6x/Gt6xMod.java
@@ -7,7 +7,7 @@ import org.altadoon.gt6x.common.items.MultiItemsX;
 import org.altadoon.gt6x.features.GT6XFeature;
 import org.altadoon.gt6x.features.basicchem.BasicChemistry;
 import org.altadoon.gt6x.features.ceramics.Ceramics;
-import org.altadoon.gt6x.features.meotorite.Meteor;
+import org.altadoon.gt6x.features.worldgen.WorldGen;
 import org.altadoon.gt6x.features.oil.OilProcessing;
 import org.altadoon.gt6x.features.pgm.PgmProcessing;
 import org.altadoon.gt6x.features.metallurgy.Metallurgy;
@@ -58,7 +58,7 @@ public final class Gt6xMod extends gregapi.api.Abstract_Mod {
 		OilProcessing.class,
 		PgmProcessing.class,
 		RefractoryMetals.class,
-		VerticalMixers.class, Meteor.class
+		VerticalMixers.class, WorldGen.class
 	};
 	private final ArrayList<GT6XFeature> enabledFeatures;
 
@@ -67,7 +67,23 @@ public final class Gt6xMod extends gregapi.api.Abstract_Mod {
 
 		this.modConfig = new Config(allFeatures);
 		this.enabledFeatures = modConfig.getEnabledFeatures();
-
+		// try to use shady reflection code to detect feature classes
+		// sometimes work sometimes dosent
+//		try{
+//			Field field =  ClassLoader.class.getDeclaredField("classes");
+//			field.setAccessible(true);
+//			Collection<Class<?>> clzs = (Collection<Class<?>>) field.get(this.getClass().getClassLoader());
+//			for(var claz:clzs){
+//				if(claz.getName().contains("org.altadoon.gt6x.features.")){
+//					System.out.println(claz.getName());
+//					if(claz.getSuperclass().equals(GT6XFeature.class)){
+//						allFeatures.add((Class<? extends GT6XFeature>) claz);
+//					}
+//				}
+//			}
+//		} catch (Exception ignored){
+//			ignored.printStackTrace();
+//		}
 		final Gt6xMod copy = this;
 		GT.mBeforePreInit.add(copy::prePreInit);
 		GT.mAfterPreInit.add(copy::postPreInit);
@@ -81,7 +97,6 @@ public final class Gt6xMod extends gregapi.api.Abstract_Mod {
 		MTx.touch();
 		RMx.init();
 		MultiItemsX.instance = new MultiItemsX(MOD_ID, "gt6x.multiitems");
-
 		for (GT6XFeature feature : enabledFeatures) {
 			feature.preInit();
 		}
